back/requirements.txt
Django
djangorestframework
django-cors-headers
django-extensions
django-filter
django-environ
django-allauth
dj-rest-auth[with_social]
psycopg2-binary
gunicorn
ipython
